+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
+## [0.2.0] - 2022-03-10
+
 - Remove `i128` feature. `Atomic{I,U}128` are now always enabled.
 
 - Add `outline-atomics` feature. Currently, this is the same as the 0.1's `i128-dynamic`, except that `fallback` feature is not implicitly enabled.
@@ -40,7 +42,8 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 Initial release
 
-[Unreleased]: https://github.com/taiki-e/portable-atomic/compare/v0.1.4...HEAD
+[Unreleased]: https://github.com/taiki-e/portable-atomic/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/taiki-e/portable-atomic/compare/v0.1.4...v0.2.0
 [0.1.4]: https://github.com/taiki-e/portable-atomic/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/taiki-e/portable-atomic/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/taiki-e/portable-atomic/compare/v0.1.1...v0.1.2